requirements.txt
pyyaml
pyyaml-include
httpx
paho-mqtt

